Renamed a package to internal in the network binding. See openhab/static-code-analysis#100
